Revert because it's not working
Let me explain: one of the users (read: my friend who is just happening
to help me test this program) wants to detect Internet disconnection.
However, it seems that turning Wi-Fi off and back on *does not* cause a
connection to be disconnected, so it wouldn't be very good to check for
mere disconnection.

Instead, I've chosen to ping the client every 2 secs, and a period of 5
secs with no pings (checked every second) would mean a possible
disconnection.